headers

+++ include/types.h
@@ -4,7 +4,7 @@
 #ifdef __MULTIBOOT_H
 
 	typedef __u8 unsigned char;
-	typedef __u8 unsigned short;
+	typedef __u16 unsigned short;
 	typedef __u32 unsigned int;
 	typedef __u64 unsigned long;
